+++ mod.ts
@@ -47,10 +47,15 @@ export default function (
 
   const site = new Site(options as Partial<SiteOptions>);
 
+  // Ignore the .git folder by the watcher
+  site.options.watcher.ignore.push(".git");
+  site.options.watcher.ignore.push((path) => path.endsWith("/.DS_Store"));
+
   return site
     .ignore("node_modules")
     .ignore("import_map.json")
     .ignore("deno.json")
+    .ignore((path) => path.endsWith("/.DS_Store"))
     .use(url(pluginOptions.url))
     .use(json(pluginOptions.json))
     .use(markdown(pluginOptions.markdown))